+++ atlas-jenkins-pipeline-gradle/gradle/wrapper/gradle-wrapper.properties
