+++ atlas-jenkins-pipeline-gradle/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'atlas-jenkins-pipeline-gradle'
